# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ascension / لغتي — Syrian-first English learning platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Profile ────────────────────────────────────────────────────────────────
...
  /exercises/{exerciseId}/submit:
    post:
      operationId: submitExercise
      tags: [exercises]
      summary: Submit answer for an MCQ or predefined exercise
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExerciseSubmission"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Exercise not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Vocabulary ─────────────────────────────────────────────────────────────
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get platform-wide admin statistics
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden (not admin)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/students:
    get:
      operationId: getAdminStudents
      tags: [admin]
      summary: List all students (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStudentList"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/lessons:
    get:
      operationId: getAdminLessons
      tags: [admin]
      summary: List all lessons (admin)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminLesson"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAdminLesson
      tags: [admin]
      summary: Create a new lesson (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLessonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLesson"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/lessons/{lessonId}:
    patch:
      operationId: updateAdminLesson
      tags: [admin]
      summary: Update a lesson (admin)
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLessonUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLesson"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
        preferredLanguage:
          type: string
          enum: [ar, en]
          default: ar
        country:
          type: string
          default: SY
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        preferredLanguage:
          type: string
          enum: [ar, en]
        country:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthResult:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # ─── Profile ───────────────────────────────────────────────────────────────
...
    ExerciseSubmission:
      type: object
      required: [selectedOptionId]
      properties:
        selectedOptionId:
          type: string
    ExerciseResult:
      type: object
      required: [correct, correctOptionId, explanation]
      properties:
        correct:
          type: boolean
        correctOptionId:
          type: string
        explanation:
          type: ["string", "null"]
        explanationAr:
          type: ["string", "null"]

    # ─── Vocabulary ────────────────────────────────────────────────────────────
...
    AdminStats:
      type: object
      required: [totalStudents, activeSubscriptions, lessonsCompleted, avgScore]
      properties:
        totalStudents:
          type: number
        activeSubscriptions:
          type: number
        lessonsCompleted:
          type: number
        avgScore:
          type: number
        generalPlanCount:
          type: number
        professionalPlanCount:
          type: number
        newStudentsThisWeek:
          type: number
    AdminStudentList:
      type: object
      required: [students, total, page, limit]
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
    AdminLesson:
      type: object
      required: [id, levelId, title, titleAr, order, lessonType, estimatedMinutes, isPublished]
      properties:
        id:
          type: number
        levelId:
          type: number
        title:
          type: string
        titleAr:
          type: string
        description:
          type: ["string", "null"]
        descriptionAr:
          type: ["string", "null"]
        order:
          type: number
        lessonType:
          type: string
          enum: [general, reading, pronunciation, speaking, vocabulary, grammar, conversation]
        estimatedMinutes:
          type: number
        isPublished:
          type: boolean
        xpReward:
          type: number
        passingScore:
          type: number
    AdminLessonInput:
      type: object
      required: [levelId, title, titleAr, lessonType, estimatedMinutes, order]
      properties:
        levelId:
          type: number
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        order:
          type: number
        lessonType:
          type: string
          enum: [general, reading, pronunciation, speaking, vocabulary, grammar, conversation]
        estimatedMinutes:
          type: number
        isPublished:
          type: boolean
        xpReward:
          type: number
        passingScore:
          type: number
    AdminLessonUpdate:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        order:
          type: number
        lessonType:
          type: string
          enum: [general, reading, pronunciation, speaking, vocabulary, grammar, conversation]
        estimatedMinutes:
          type: number
        isPublished:
          type: boolean
        xpReward:
          type: number
        passingScore:
          type: number