# lib/api-zod/src/generated/types/adminLessonUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ascension / لغتي — Syrian-first English learning platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminLessonUpdateLessonType } from './adminLessonUpdateLessonType';

export interface AdminLessonUpdate {
  title?: string;
  titleAr?: string;
  description?: string;
  descriptionAr?: string;
  order?: number;
  lessonType?: AdminLessonUpdateLessonType;
  estimatedMinutes?: number;
  isPublished?: boolean;
  xpReward?: number;
  passingScore?: number;
}