# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Ascension / لغتي — Syrian-first English learning platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPreferredLanguage } from './userPreferredLanguage';
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  preferredLanguage?: UserPreferredLanguage;
  country?: string;
  createdAt: Date;
}